Fix uniform work distributor.
The algorithm required the candidates list to be sorted from the lowest id to
the highest.

Actually the algorithm needs to have a stable order in which the assignation
must take place.  We keep the last ID assigned and then compare that to
candidates and select the next biggest id or wrap around.

The previous algorithm used `min` to select the lowest possible id.  But
since, now we ensure that ids are ordered with simply select the first that
bigger than the last selected candidate.