Add a new Value mixin to hold values for any identifier.
The CDR Identifier (used by control variables and evidences) holds a
Python value that is stored as a VARCHAR in the DB.

That type of value is also used in the history of values of those
objects.

Sentry Issue [1] reveals that 'value' should always be regarded as the
repr'ed value.

To avoid some duplication, a single Value mixin holds both the `value`
field and a new `result` Property.

`fields.Property` was first developed here, but soon after migrated to
`xoeuf.fields`; that's why you don't see any trace of its
implementation here.

Solves: MERCURIO-1A8

.. [1] http://sentry.lahavane.com/caraibes/mercurio/issues/4949/