Reimplement object merger.
There's now a public API for the merger.  See the `merge` method.

The new implementation removes the general merge method that used Python code.
Now a fixed set of merge methods are available.

Remove a previously-undeclared dependency of the object merge with the addon
'mail'.  This dependency is now modeled with an integration addon
'xopgi_i9n_objectmerger_mail'.

Add tests for the merge.