Don't check for cycles.
Cycles may be allowed per model.  There's nothing wrong with cycles per se.
What we should do it to check the constrains of the model.  If the model does
not allow checks it will raise ValidationError if we're creating one.

TODO: Due to our current algorithm the same initial state may or may not
create a cycle.  Given

     A <----- B <------ C

if we merge A and C, and select A as target we get

     A <----- B

but if we select C as target we get

              B <------ C
              |         ^
              +---------+