Extract description level out of the main model.
This is a big model.  There are clearly two parts of it:

- The description of a recurrence rule.

- The computation of occurrences of a recurrence rule.

I'm planning to move to the module 'describe.py' all things related to
description.  This is human-oriented descriptions which allows nice user
interfaces.

I'm doing it in abstracts models that I will mix into our current
'recurrent.model'.  This maintains the same DB layout but sets the path to
more physical divisions.