Use 'recurrent.model' as a mixin in CDR.
We still keep the definition of the recurrence in a separate model:
'cdr.recurrent.event.def'.  And make 'cdr.recurrent.event' to *_inherits* from
it.

The 'recurrence' fields is converted to a Property that aliases the new
relation.

Migrations are provided that convert all current 'recurrent.model' instances
which are related to CDR events into 'cdr.recurrent.event.def' instances.