Add method to obtain all occurrences.
I'm targeting to rewrite this code in a functional-like fashion.  The current
implementation can be hardly limited because each recurrent definition is
processed at once and limited (to avoid infinite occurrences).

This is bad because given two recurrent definitions that produce the following
pattern of occurrences:

   |---*---*---*---*---*---* .... and so on

   |-----x-----x-----x-----x .... and so on

Since both are unlimited the current searching method will only return items
from the first recurrence.

Using a `merged` iterator that yields the next-in-order would produce a single
iterator with the desired mixing of items:

   |---*-x-*---*x---*-x-*---*x .... and son

(Note that '*x' are given in that order but they co-occur.)

This method (with the help of Predicate) will replace the current searching
procedure.